+++ src/store/UserModule.js
@@ -1,4 +1,5 @@
 import * as userApi from "@/api/user-api";
+import * as downloadApi from "@/api/download-api";
 
 export default {
   namespaced: true,
@@ -15,6 +16,15 @@ export default {
   actions: {
     async getUsers(context) {
       const response = await userApi.getUsers();
+
+      await Promise.all(
+        response.map(async (user) => {
+          if (user.avatar) {
+            user.avatarUrl = await downloadApi.getDownloadAsUrl(user.avatar.id);
+          }
+        }),
+      );
+
       context.commit("setUsers", response);
     },
   },